Move `SearchLink` class to `views/logic/search`.
