+++ graylog2-web-interface/src/views/logic/search/SearchLink.test.tsx
